+++ RELEASE NOTES.md
@@ -14,6 +14,8 @@ MVC 6 API
 - Upgraded NWebSec from 1.0.0 to 1.1.0.
 - Run health checks in parallel in StatusController.
 - Add Last-Modified and If-Modified-Since HTTP headers to return 304 Not Modified if a car has not changed.
+- Response caching is now an optional feature.
+- Upgraded all NuGet packages.
 
 ### 1.5.0
 MVC 6 API