+++ RELEASE NOTES.md
@@ -8,6 +8,9 @@
 
 # dotnet new Boilerplate.Templates NuGet Package
 
+### 1.6.1
+- Use AppendCommaSeparatedValues and HeaderNames constants.
+
 ### 1.6.0
 ASP.NET Core API Boilerplate
 - Added Prefix.io profiler option.
@@ -16,7 +19,6 @@ ASP.NET Core API Boilerplate
 - Add Last-Modified and If-Modified-Since HTTP headers to return 304 Not Modified if a car has not changed.
 - Response caching is now an optional feature.
 - Upgraded all NuGet packages.
-- Use AppendCommaSeparatedValues and HeaderNames constants.
 
 ### 1.5.0
 ASP.NET Core API Boilerplate